tarea finalizada

+++ src/hooks/UseFormula.js
@@ -18,9 +18,9 @@ const Formula = () =>  {
        } else if (ValorC === '') {
             Alerta('Ingrese el valor de C')
     }else{
-    let discriminante = Math.pow(parseFloat(ValorB),2) - 4 * parseFloat(ValorA) * parseFloat(ValorC);
-    let x1 = ((-parseFloat(ValorB) + Math.sqrt(discriminante)) / (2 * parseFloat(ValorA))); 
-    let x2 = ((-parseFloat(ValorB) - Math.sqrt(discriminante)) / (2 * parseFloat(ValorA)));
+    let discriminante = Math.pow(parseInt(ValorB),2) - 4 * parseInt(ValorA) * parseInt(ValorC);
+    let x1 = ((-parseInt(ValorB) + Math.sqrt(discriminante)) / (2 * parseInt(ValorA))); 
+    let x2 = ((-parseInt(ValorB) - Math.sqrt(discriminante)) / (2 * parseInt(ValorA)));
     setResultado(x1);
     setResultado2(x2);
     }
